# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediaVault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/folders:
    get:
      operationId: getLibraryFolders
      tags: [library]
      summary: List all library folders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryFolder"
    post:
      operationId: addLibraryFolder
      tags: [library]
      summary: Add a library folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryFolderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolder"
  /library/folders/{id}:
    delete:
      operationId: removeLibraryFolder
      tags: [library]
      summary: Remove a library folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /library/scan:
    post:
      operationId: scanLibrary
      tags: [library]
      summary: Trigger a library scan
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanOptions"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"

  # ── Media Files ───────────────────────────────────────────────────────────
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List and search media files
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [video, audio, all]
        - name: tags
          in: query
          schema:
            type: string
        - name: rating
          in: query
          schema:
            type: integer
        - name: favorite
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
            enum: [available, offline, all]
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: filters
          in: query
          schema:
            type: string
          description: JSON-encoded advanced filter object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get a single media file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "404":
          description: Not found
    patch:
      operationId: updateMedia
      tags: [media]
      summary: Update media metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Remove a media file from the library (does not delete the file)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /media/{id}/metadata:
    get:
      operationId: getMediaMetadata
      tags: [media]
      summary: Get all metadata for a media file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaMetadataEntry"
    put:
      operationId: setMediaMetadata
      tags: [media]
      summary: Replace all metadata for a media file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaMetadataInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaMetadataEntry"
  /media/bulk:
    patch:
      operationId: bulkUpdateMedia
      tags: [media]
      summary: Bulk update media metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMediaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResult"

  # ── Metadata Categories & Values ─────────────────────────────────────────
...
components:
  schemas:
...
    LibraryFolder:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        lastScanned:
          type: ["string", "null"]
        fileCount:
          type: integer
        createdAt:
          type: string
      required: [id, path, name, enabled, fileCount, createdAt]
    LibraryFolderInput:
      type: object
      properties:
        path:
          type: string
        name:
          type: string
      required: [path]
    ScanOptions:
      type: object
      properties:
        folderId:
          type: ["integer", "null"]
        full:
          type: boolean
    ScanResult:
      type: object
      properties:
        added:
          type: integer
        updated:
          type: integer
        removed:
          type: integer
        duration:
          type: number
      required: [added, updated, removed, duration]
    MediaFile:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [video, audio]
        size:
          type: integer
        duration:
          type: ["number", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        fps:
          type: ["number", "null"]
        bitrate:
          type: ["integer", "null"]
        codec:
          type: ["string", "null"]
        format:
          type: string
        thumbnail:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        favorite:
          type: boolean
        playCount:
          type: integer
        lastPlayed:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, offline]
        dateAdded:
          type: string
        dateModified:
          type: string
        folderId:
          type: ["integer", "null"]
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/MediaMetadataEntry"
      required: [id, path, filename, type, size, format, favorite, playCount, status, dateAdded, dateModified]
    MediaPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaFile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
      required: [items, total, page, limit, pages]
    MediaUpdate:
      type: object
      properties:
        rating:
          type: ["integer", "null"]
        favorite:
          type: boolean
        thumbnail:
          type: ["string", "null"]
    MediaMetadataEntry:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        valueId:
          type: integer
        value:
          type: string
      required: [categoryId, categoryName, categorySlug, valueId, value]
    MediaMetadataInput:
      type: object
      properties:
        entries:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: integer
              value:
                type: string
            required: [categoryId, value]
      required: [entries]
    BulkMediaUpdate:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        rating:
          type: ["integer", "null"]
        favorite:
          type: boolean
        addMetadata:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: integer
              value:
                type: string
            required: [categoryId, value]
        removeMetadataCategoryIds:
          type: array
          items:
            type: integer
      required: [ids]
    BulkResult:
      type: object
      properties:
        updated:
          type: integer
      required: [updated]
...