# lib/api-zod/src/generated/types/mediaMetadataEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaVault API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaMetadataEntry {
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  valueId: number;
  value: string;
}