# lib/api-zod/src/generated/types/mediaMetadataInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaVault API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaMetadataInputEntriesItem } from './mediaMetadataInputEntriesItem';

export interface MediaMetadataInput {
  entries: MediaMetadataInputEntriesItem[];
}