# lib/api-zod/src/generated/types/mediaPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaVault API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFile } from './mediaFile';

export interface MediaPage {
  items: MediaFile[];
  total: number;
  page: number;
  limit: number;
  pages: number;
}